Break out Http request code + fixup phpunit + composer autload
